Restructure remote
